requirements.txt
anthropic
python-dotenv
PyMuPDF
python-docx
